+++ server/src/main/java/com/backend/controller/dto/post/CreatePostRequest.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 public class CreatePostRequest {
   private String content;
-  private Optional<Long> rootPostId = Optional.empty();
-  private Optional<Long> postId = Optional.empty();
+  private Optional<Integer> rootPostId = Optional.empty();
+  private Optional<Integer> postId = Optional.empty();
 }
