+++ server/src/main/java/com/backend/service/UserService.java
@@ -50,7 +50,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User getUserById(Long id) {
+    public User getUserById(Integer id) {
         return userRepository.findById(id).get();
     }
